moved seed function to seed.js and added Order/OrderDrink associations to server/db/index.js
